[PROD-8246]: Ability to use Okta for end user login
Add okta-spring-boot-starter dep
Add Identity provider in CsmPlatformProperties
Extract roles/paths mapping and configuration in AbstractSecurityConfiguration
Separate CsmAzureSecurityConfiguration and OktaSecurityConfiguration
Add possibility to overwrite roles/scope prefixes in CsmAzureBasedSecurityConfiguration
